requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
gspread==5.10.0
google-auth==2.22.0
yfinance==0.2.20
